files committed
I analyzed CSV data and created an interface using MATLAB. Based on the information entered into this interface, I compared it with the information obtained from the CSV data to calculate the risk of heart disease. This project aims to provide a predictive analysis of heart disease risk, contributing to early detection and prevention strategies.